feat(genai): Update samples to use Gemini 2.5 Flash MM and separate test to use global endpoint.
