+++ genai/image_generation/imggen_mmflash_edit_img_with_txt_img.py
@@ -26,7 +26,7 @@ def generate_content() -> str:
     image = Image.open("test_resources/example-image-eiffel-tower.png")
 
     response = client.models.generate_content(
-        model="gemini-2.0-flash-preview-image-generation",
+        model="gemini-2.5-flash-image-preview",
         contents=[image, "Edit this image to make it look like a cartoon."],
         config=GenerateContentConfig(response_modalities=[Modality.TEXT, Modality.IMAGE]),
     )
